+++ java-dynamic-sqs-listener-spring/java-dynamic-sqs-listener-spring-starter/src/main/java/com/jashmore/sqs/spring/container/batching/BatchingQueueListener.java
@@ -80,7 +80,7 @@ public @interface BatchingQueueListener {
      * @return the period in ms that threads will wait for messages to be requested from SQS
      * @see BatchingMessageRetrieverProperties#getMessageRetrievalPollingPeriodInMs() for more details
      */
-    int maxPeriodBetweenBatchesInMs() default 2000;
+    long maxPeriodBetweenBatchesInMs() default 2000L;
 
     /**
      * The message visibility that will be used for messages obtained from the queue.
